Release 3.3.1
- Allow input files with multiple channels (mixed input file)
- New option customApi for CLI version
- Malformed utf8 detection
- Improved detection system for malformed records
- Minor improvements in SimpleApiSimulator
- Warn when normalized record is empty
- Fix: <could not open file> error with very large files
- Fix: create outputFolder if not exists (server)
- More tests

+++ src/cli/master.js
@@ -44,7 +44,7 @@ const workerReport = (workerMsg) => {
       }
 
       taskToWorker(workers[workerMsg.workerId-1], workerParams, params.linesPerTask)
-        .catch(() => {throw new Error('Fatal error feeding worker')})
+        .catch((e) => finish(5, 'Fatal error feeding worker: ' + e))
       break
   }
 }
@@ -68,18 +68,25 @@ const requestLines = async (lineGenerator, numberOfLines) => {
   return lines
 }
 
+const finishFiles = async () => {
+  if (success) {
+    await success.save(params.outputFolder)
+    success.delete()
+  }
+  if (fail) {
+    await fail.save(params.outputFolder)
+    fail.delete()
+  }
+}
 const finish = async (code, errMsg) => {
   try {
     clearInterval(reportTimer)
-    if (errMsg) process.stderr.write(errMsg)
-    await success.save(params.outputFolder)
-    await fail.save(params.outputFolder)
-    success.delete()
-    fail.delete()
+    if (errMsg) process.stderr.write(errMsg + '\n')
+    if (success || fail) await finishFiles()
     report.printEnd(code)
     process.exit(code)
   } catch(e) {
-    process.stderr.write('\nError cleaning session. Some results may have been lost.')
+    process.stderr.write('\nError cleaning session. Some results may have been lost. Error:' + e)
     process.exit(1)
   }
 }
@@ -93,7 +100,7 @@ const workerEnd = (worker, code, signal) => {
     const errMsg =
       "Fatal error (" + (signal || code) + "): a child process ended with errors. " +
       "Try again reducing current linesPerTask ("+ params.linesPerTask + ") or " +
-      "current number of cores to use (" + params.coresToUse + ").\n"
+      "current number of cores to use (" + params.coresToUse + ")."
     killWorkers()
     finish(2, errMsg)
   } else {
@@ -113,19 +120,20 @@ const setupMaster = async () => {
   //Setup master
   lineGenerator = readLines(params.inputCsv)
 
-  success = new HugeFile(params.inputCsv.name, '_slr-client-success_', 0, fs, params.outputFolder)
-  fail = new HugeFile(params.inputCsv.name, '_slr-client-error_', 0, fs, params.outputFolder)
-
   workerParams.inputCsv = params.inputCsv
   workerParams.channel = params.channel
   workerParams.credentials = { key: params.key, secret: params.secret }
   workerParams.config = config
 
-  params.totalRecords = await countLines(params.inputCsv)
-  if (params.totalRecords < params.linesPerTask*params.coresToUse)
-    params.linesPerTask = Math.ceil(params.totalRecords/params.coresToUse)
   consolidatedProgress = new ConsolidatedProgress(params.totalRecords)
   report = new Report(consolidatedProgress, params.checkUpdates, config.latestVersionCheckUrl, packageJson.version)
+
+  try { params.totalRecords = await countLines(params.inputCsv) }
+  catch(e) { await finish(3, e.message) }
+  consolidatedProgress.updateTotalRecords(params.totalRecords)
+
+  if (params.totalRecords < params.linesPerTask*params.coresToUse)
+    params.linesPerTask = Math.ceil(params.totalRecords/params.coresToUse)
 
   process.on("SIGINT", () => {
     cluster.off('exit', workerEnd)
@@ -134,10 +142,15 @@ const setupMaster = async () => {
   })
   cluster.on('exit', workerEnd)
 
+  fs.mkdirSync(params.outputFolder, {recursive: true})
+  success = new HugeFile(params.inputCsv.name, '_slr-client-success_', 0, fs, params.outputFolder)
+  fail = new HugeFile(params.inputCsv.name, '_slr-client-error_', 0, fs, params.outputFolder)
+
   //Create workers and send tasks
   for (let i = 0; i < params.coresToUse; i++) workers.push(cluster.fork())
   for (let w of workers) {
     await taskToWorker(w, workerParams, params.linesPerTask)
+      .catch((e) => finish(4, 'Fatal error feeding worker for the first time: ' + e))
     w.on('message', (msg) => {
       workerReport(msg)
     })
